requirements.txt
sentence-transformers
faiss-cpu
requests
numpy
fastapi
uvicorn
psycopg2
email-validator
dotenv
python-multipart
streamlit